Updated to 1.6.2-R0.1
Had to update dependencies (ugh!)

+++ src/plugin/bleachisback/LogiBlocks/Commands/SetHealthCommand.java
@@ -30,7 +30,7 @@ public class SetHealthCommand extends BaseCommand
 			sender.sendMessage(ChatColor.RED+"That's not living!");
 			return true;
 		}
-		int setHp=0;
+		double setHp=0.0;
 		try
 		{
 			switch(args[1].charAt(0))
